requirements-dev.txt
ansible>=2.10.0
ansible-lint>=6.0.0
molecule>=4.0.0
molecule-docker>=2.0.0
pytest-testinfra>=6.0.0
docker>=6.0.0

